+++ Exe/Pdbstrr.fs
@@ -1,6 +1,0 @@
-﻿module SourceLink.Pdbstrr
-
-let run (pdb: string) =
-    use p = new PdbFile(pdb)
-    for line in p.ReadSrcSrvLines() do
-        printfn "%s" line