+++ SymbolStore/HResult.fs
@@ -1,4 +1,0 @@
-﻿[<AutoOpen>]
-module SourceLink.SymbolStore.HResult
-
-[<Literal>] let S_OK = 0