+++ lib/Copy-Libraries.ps1
@@ -1,4 +1,0 @@
-copy 'C:\Program Files (x86)\MSBuild\14.0\Bin\Microsoft.Build.dll' $PSScriptRoot
-copy 'C:\Program Files (x86)\MSBuild\14.0\Bin\Microsoft.Build.Framework.dll' $PSScriptRoot
-copy 'C:\Program Files (x86)\MSBuild\14.0\Bin\Microsoft.Build.Utilities.Core.dll' $PSScriptRoot
-copy 'C:\Windows\Microsoft.NET\assembly\GAC_MSIL\System.Threading.Tasks.Dataflow\v4.0_4.5.24.0__b03f5f7f11d50a3a\System.Threading.Tasks.Dataflow.dll' $PSScriptRoot